continuing work on 3 class u-net

+++ u-net.py
@@ -473,7 +473,7 @@ def main(mode,
             )
             probs_summary = tf.reduce_max(
                 tf.multiply(
-                    max_bool,
+                    tf.cast(max_bool,tf.float32),
                     tf.stack([1 - prediction_summary[:,:,:,0],
                               prediction_summary[:,:,:,1],
                               1 - prediction_summary[:,:,:,2]])
